Separate Data namespace to DataType and Interface
