+++ Assets/UnitySensors/Editor/DataType.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c32a713a080943347936bc27971033d7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
