+++ Assets/UnitySensors/Editor/DataType/LiDAR.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ada79c3060577e74d986fcd3d8a5b8a8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
