--- Assets/UnitySensors/Editor/LiDAR.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4967c6c40eb14a846b6c0697633a4ac4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
