--- Assets/UnitySensors/Runtime/Scripts/Data/PointCloud/IPointCloudInterface.cs
@@ -1,10 +1,0 @@
-using Unity.Collections;
-
-namespace UnitySensors.Data.PointCloud
-{
-    public interface IPointCloudInterface<T> where T : struct, IPointInterface
-    {
-        public PointCloud<T> pointCloud { get; }
-        public int pointsNum { get; }
-    }
-}
