--- Assets/UnitySensors/Runtime/Scripts/Data/PointCloud/IPointCloudInterface.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 004afa2710acbb34b9ce154b1ce02213
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
