--- Assets/UnitySensors/Runtime/Scripts/Data/PointCloud/IPointInterface.cs
@@ -1,9 +1,0 @@
-using Unity.Collections;
-
-namespace UnitySensors.Data.PointCloud
-{
-    public interface IPointInterface
-    {
-        public void CopyTo(NativeArray<byte> dst);
-    }
-}
