--- Assets/UnitySensors/Runtime/Scripts/Data/PointCloud/IPointInterface.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1c41a96fc5b816a4caf653c7245c7be7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
