--- Assets/UnitySensors/Runtime/Scripts/Data/PointCloud/IPointXYZInterface.cs
@@ -1,9 +1,0 @@
-using Unity.Mathematics;
-
-namespace UnitySensors.Data.PointCloud
-{
-    public interface IPointXYZInterface : IPointInterface
-    {
-        public float3 position { get; }
-    }
-}
