--- Assets/UnitySensors/Runtime/Scripts/Data/PointCloud/IPointXYZInterface.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5f80a0f4e6ff25e4982e0c6d7b22666e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
