--- Assets/UnitySensors/Runtime/Scripts/Data/Pose.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 178d5051c9466c44c8cb92fb4a78d8ef
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
