--- Assets/UnitySensors/Runtime/Scripts/Data/Pose/IPoseInterface.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-namespace UnitySensors.Data.Pose
-{
-    public interface IPoseInterface
-    {
-        public Vector3 position { get; }
-        public Quaternion rotation { get; }
-    }
-}
