--- Assets/UnitySensors/Runtime/Scripts/Data/Texture.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9accbaf0f228f9b43bc848e37c637881
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
