--- Assets/UnitySensors/Runtime/Scripts/Data/Texture/ITextureInterface.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace UnitySensors.Data.Texture
-{
-    public interface ITextureInterface
-    {
-        public Texture2D texture { get; }
-    }
-}
