--- Assets/UnitySensors/Runtime/Scripts/Data/Texture/ITextureInterface.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 733a3f55ec218c04f95cc796ed3926fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
