+++ Assets/UnitySensors/Runtime/Scripts/DataType/Geometry.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5a9ed2b29176c864f88348b83347109c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
