+++ Assets/UnitySensors/Runtime/Scripts/DataType/Geometry/GeoCoordinate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6e09cfe85ee8c746933b974567f627e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
