+++ Assets/UnitySensors/Runtime/Scripts/DataType/Geometry/Vector3D.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 678783905e1a14645a14f3dff365ba19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
