+++ Assets/UnitySensors/Runtime/Scripts/DataType/LiDAR.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d71673b62fc54b7429dcabeb6e63e493
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
