+++ Assets/UnitySensors/Runtime/Scripts/DataType/LiDAR/ScanPattern.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92acdf05f4e317341ab30489744293dd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
