+++ Assets/UnitySensors/Runtime/Scripts/DataType/Sensor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fd7f6539d113f084b90db0a95172558e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
