+++ Assets/UnitySensors/Runtime/Scripts/DataType/Sensor/PointCloud/PointXYZRGB.cs
@@ -1,0 +1,23 @@
+using System;
+using Unity.Collections;
+using Unity.Mathematics;
+
+using UnitySensors.Interface.Sensor.PointCloud;
+
+using UnityEngine;
+
+namespace UnitySensors.DataType.Sensor.PointCloud
+{
+    [System.Serializable]
+    public struct PointXYZRGB : IPointInterface
+    {
+        [SerializeField]
+        private float3 _position;
+        public byte b;
+        public byte g;
+        public byte r;
+        public byte a;
+
+        public float3 position { get => _position; set => _position = value; }
+    }
+}
