+++ Assets/UnitySensors/Runtime/Scripts/DataType/Sensor/PointCloud/PointXYZRGB.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49f836b5b85e31243a6a556eacb3e347
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
