+++ Assets/UnitySensors/Runtime/Scripts/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b82e81a19ae443c4a831ab5a4aebfacf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
