+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Geometry.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3195f74ad8ce91e4fb74ba3474ceaa1b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
