+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Geometry/IGeoCoordinateInterface.cs
@@ -1,0 +1,9 @@
+using UnitySensors.DataType.Geometry;
+
+namespace UnitySensors.Interface.Geometry
+{
+    public interface IGeoCoordinateInterface
+    {
+        public GeoCoordinate coordinate { get; }
+    }
+}
