+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Geometry/IGeoCoordinateInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0e2ae104118e9349a42baa4333c1998
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
