+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Geometry/IPoseInterface.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace UnitySensors.Interface.Geometry
+{
+    public interface IPoseInterface
+    {
+        public Vector3 position { get; }
+        public Quaternion rotation { get; }
+    }
+}
