+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 162b3d58085a39f4f91a44a79c74a006
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
