+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/ICameraInterface.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace UnitySensors.Interface.Sensor
+{
+    public interface ICameraInterface
+    {
+        public Camera m_camera { get; }
+    }
+}
