+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/ICameraInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf69d703a2e42b844b726a16214b80b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
