+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/IImuDataInterface.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace UnitySensors.Interface.Sensor
+{
+    public interface IImuDataInterface
+    {
+        public Vector3 acceleration { get; }
+        public Quaternion rotation { get; }
+        public Vector3 angularVelocity { get; }
+    }
+}
