+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/IImuDataInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6792853748668214798c99d9bab79841
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
