+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/IPointCloudInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93051d61501a2f94ca8fb9aa37c8bfd4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
