+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/ITextureInterface.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace UnitySensors.Interface.Sensor
+{
+    public interface ITextureInterface
+    {
+        public Texture2D texture0 { get; }
+        public Texture2D texture1 { get; }
+    }
+}
