+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/ITextureInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41dfa054de7d83d4ba0b91d5a7d3ef0e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
