+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/PointCloud.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d5a7872469422c4e8ec1caeff004a5c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
