+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/PointCloud/IPointInterface.cs
@@ -1,0 +1,10 @@
+using Unity.Collections;
+using Unity.Mathematics;
+
+namespace UnitySensors.Interface.Sensor.PointCloud
+{
+    public interface IPointInterface
+    {
+        public float3 position { get; set; }
+    }
+}
