+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Sensor/PointCloud/IPointInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1023061985d9df459b16422f5e68670
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
