+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Std.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 612243489ad8c244eba80780a37bffe1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
