+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Std/ITimeInterface.cs
@@ -1,0 +1,7 @@
+namespace UnitySensors.Interface.Std
+{
+    public interface ITimeInterface
+    {
+        public float time { get; }
+    }
+}
