+++ Assets/UnitySensors/Runtime/Scripts/Interfaces/Std/ITimeInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9bb309144c8ff640ae0afa217d3262f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
